Infra/terraform/modules/databricks_dashboard_content/requirements.txt
azure-identity==1.18.0
requests==2.32.3
